fix: moving files around in client
